update for zlib & FreeImage
git-svn-id: http://svn.miranda-ng.org/main/trunk@238 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/FreeImage/Source/FreeImage/BitmapAccess.cpp
@@ -169,7 +169,7 @@ FreeImage_GetImageSizeHeader(BOOL header_only, unsigned width, unsigned height, 
 		const size_t header_size = dib_size;
 
 		// pixels are aligned on a 16 bytes boundary
-		dib_size += CalculatePitch(CalculateLine(width, bpp)) * height; 
+		dib_size += (size_t)CalculatePitch(CalculateLine(width, bpp)) * (size_t)height; 
 
 		// check for possible malloc overflow using a KISS integer overflow detection mechanism
 		{
